Put stripPreReleaseWarning() next to withPreamble()
