+++ test/solc/Common.cpp
@@ -17,6 +17,7 @@
 // SPDX-License-Identifier: GPL-3.0
 
 #include <test/solc/Common.h>
+#include <test/libsolidity/util/Common.h>
 
 #include <solc/CommandLineInterface.h>
 
@@ -78,18 +79,4 @@ test::OptionsReaderAndMessages test::runCLI(
 		sout.str(),
 		stripPreReleaseWarning(serr.str()),
 	};
-}
-
-string test::stripPreReleaseWarning(string const& _stderrContent)
-{
-	static regex const preReleaseWarningRegex{
-		R"(Warning( \(3805\))?: This is a pre-release compiler version, please do not use it in production\.\n)"
-		R"((\n)?)"
-	};
-	static regex const noOutputRegex{
-		R"(Compiler run successful, no output requested\.\n)"
-	};
-
-	string output = regex_replace(_stderrContent, preReleaseWarningRegex, "");
-	return regex_replace(std::move(output), noOutputRegex, "");
 }
